remove qusb compatibility
As far as I can tell, sni offers the same functionality as qusb
but is more stable (speaking only from own experiences) and has a better communication protocol.
The experience of installing sni on a non-windows machine (been using
ubuntu) was a breeze compared to qusb, so there's that as well.

If there ever comes a legit reason to have qusb support, then I'll
implement it in a better way than this adhoc feature toggling crap I
went with now, which was poorly designed and just annoying to keep
updated.